Refactor repository layer and service layer to follow ES6 class syntax and leverage class-based inheritance
